+++ src/db/models/publication/mod.rs
@@ -41,7 +41,7 @@ pub trait TxManager {
         &mut self,
         name: &str,
         stele: &str,
-    ) -> anyhow::Result<Publication>;
+    ) -> anyhow::Result<Option<Publication>>;
     /// Find the first publication by date and stele that is not revoked.
     async fn find_first_by_date_and_stele_non_revoked(
         &mut self,